backend/requirements.txt
# FastAPI with JWT Auth and SQLAlchemy (Updated for Python 3.13)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
# psycopg2-binary==2.9.9  # Only needed for PostgreSQL, commented out for SQLite development
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
alembic>=1.14.0
